Probability and Statistical Inference Project in R
Part 1: 

Assess the effect of a collection of explanatory variables on the plasma 
ferritin concentration (Ferr) in 202 Australian athletes. Split data into training and testing sets, create a model and evaluate it.

Part 2: Bayesian Inference

An unknown value 𝜃 has been transmitted to you over a noisy channel. Let us assume the noise is 
normally distributed with mean 0 and a known variance 4. So the value x that you receive is modelled 
by 𝑁(𝜃, 4). Based on previous communications, your prior knowledge on 𝜃 is 𝑁(12, 9).